Implement heartbeating over the RPC connection
Having to implement this ourselves on the application
layer is a bit annoying, but heartbeats are the only solution
that reliably detects cases where the OS does not know that
the TCP peer disappeared, e.g. if the network cable is removed
from the server or it is powered off abruptly.